More cost functions added and tested
The following cost functions were programmed into the "PTG" class and
tested:
- timeDiffCost: Difference between the desired time for the trajectory
and the real executed time
- sDiffCost: Difference between the s coordinates of the target car and
the s coordinates of the ego car at the end of the trajectory.
- dDiffCost: Difference between the d coordinates of the target car and
the d coordinates of the ego car at the end of the trajectory.

The timeDiffCost function was tested by generating a trajectory and
defining an end time of 5.0 seconds at the beginning and changing it in
order to see how the cost functions reacts.

The sDiffCost function was tested by generating a trajectory and
defining a target car which moves with constant acceleration. The end
parameters of the trajectory were changed in order to see how the cost
function reacts.

The dDiffCost function was tested the same way as the sDiffCost
function.