More cost functions added
The cost functions "collisionCost" and "bufferCost" were added,
including their help methods "nearestApproach" and
"nearestApproachToAnyVehicle".

The cost function "collisionCost" was tested generating trajectories
which collide with the other cars, observing the expected result, the
function returns 1 if a collision is detected.

The cost function "bufferDistCost" was tested generating trajectories
which drive dangerously close to the other cars, observing the expected
result, the function returns values close to 1 if the car was to close
to other one during the trajectory.