More cost functions added.
Cost functions "goOutRoadCost", "exceedsSpeedLimitCost" and
"efficiencyCost" were programmed and tested.

The cost function "goOutRoadCost" was tested by generating trajectories
which goes out of the road. As expected, a trajectory which runs
completely out of the road gives a cost close to 1. If the trajectory
only goes for a little while out of the road, the cost will be between 0
and 0.9 depending on how long it was out of the road and how far outside
of the road it was.

The cost function "exceedsSpeedLimitCost" was tested by generating
trajectories which exceed the speed limit. Exactly as the
"goOutRoadCost" function, the cost depends on how long the speed limit
was exceeded and how high that speed was.

The cost function "efficiencyCost" was tested by generating trajectories
which go faster or slower than the target car. As expected, if the
target car was faster on average than the ego car, the cost will be
positive and if the ego car was faster on average than the target car,
the cost will be negative.