General methods to generate best trajectory added
The following methods were added and tested:
- calculateTotalCost: Calculates the maximum trajectory cost based on
the cost functions and their weights. It was tested asigning to some
cost functions the weight 1 while the weight of the other ones stays at
0 and comparing the result with the expected result.

- "perturbGoal": Calculates noisy goal values based on the original goal
and the corresponding standard deviations. It was tested by using it
multiple times on the same goal, observing different result goals
arround this first defined goal.

- "getBestFollowTrajectory": Generates many trajectories with
goal arround the original goal and time arround the original desired
time and selects the one which has the smaller cost. This method can not
be tested so easily. In order to test it, the same case as in the
program in Python was generated using this method observing
similar costs of the best trajectory generated.