All methods of PTG class optimized
All methods of PTG class were optimized and their parameters changed in
order to be used also without a target car. In order to get a goal based
on a target, the method "getGoalBasedOnTarget" was programmed.

All was tested using exactly the same test case used on the last commit
and comparing the total cost and the individual costs of all the
functions obtaining exactly the same result.