The class "ego_vehicle" was started
The class "ego_vehicle" was started, which is responsible of controlling
the ego vehicle. It aggregates the class "PTG" which is responsible of
creating the best possible trajectory.

The following methods were already programmed and tested to the class:
- getCircularTraj: It is a simple method in order to test the
interfacing between this program and the simulator. It creates a
circular path without any planning and sends it to the simulator. With
this class, it was tested how the points of the last trajectory can be
used in the new trajectory, which is very important for any future
trajectory generation.

- getThetasFromXY: This method gets the yaws of a car based on a vector
of x values and a vector of y values. These values are needed in order
to transform a trajectory in XY coordinates to Frenet coordinates. This
method was tested getting many points in a circular trajectory generated
by "getCircularTraj", so the thetas of this trajectory are continuous
between -pi and pi. The expected result was obtained on the test.

- trajXYToFrenet and trajFrenetToXY: These methods, as the names say
convert a trajectory from XY to Frenet and from Frenet to XY
respectively. They were tested converting the circular trajectory
generated with "getCircularTraj" to Frenet using "trajXYToFrenet" and
then back to XY using "trajFrenetToXY", obtaining almost exactly the
same trajectory as the original one in XY. Only the very last points of
the trajectory are not that exact, so these points should not be used
for further calculations.