Methods of "ego_vehicle" added and tested
The following methods were added to the "ego_vehicle" class:
"getVelocity", "getAcceleration" "PTGkeepLineTraj" and "SplineTraj". The
first three methods are meant to be used generating a quintic polynomic
trajectory.

The method "getVelocity" gets a vector of velocities based on a vector
of positions on any coordinate system. It was tested using a vectors of
positions on s and d giving the right result, but also indicating that
the time distance between the points can not be right because the
resulting speed was too high. Lastly this method was changed by a way
easier solution which is saving the last speed on s and d within the
variables "current_speed_s" and "current_speed_d.

The method "getAcceleration" works the same way as "getVelocity", so it
gets an array of acceleration based on an array of velocities. It was
tested the same way and the same results were obtained. This method wa
changed by saving the last acceleration on "current_acc_s" and
"current_acc_d".

The method "PTGkeepLineTraj" generates a trajectory using the PTG class.
It works fairly well, but it causes some max jerk and max acceleration
alarms every now and then. These alarms are caused because of the
transformation between Frenet and XY coordinates which is not perfect
and generates discontinuities. A lot of things were tested in order to
eliminate those alarms but nothing worked. For now, this method will not
be used further in the project and will be replaced by "SplineTraj"
which does not rely too much on Frenet coordinates.

The method "SplineTraj" implements the starting idea presented by the
Udacity tutors on the Q&A video of the project. It generates a
trajectory on X-Y instead of two separated trajectories for t-X and t-Y,
it only uses 3 points on Frenet coordinates so it minimizes the
transformation error and generates a spline curve between the equivalent
XY of those points. It works very well not generating max jerk and max
acceleration alarms after the start. Until now these alarms only happen
at the beginning of the program, since the car speed jumps from 0 mph to
50 mph instantaneously. This issue still needs to be corrected.