More methods of "ego_vehicle" added and tested
The following methods were added to "ego_vehicle": changeTrajectory,
getVehicleAhead, getVehicleBehind and getKinematicsOfLane.

The method "changeTrajectory" is responsible of changing the lane of the
car and its velocity and acceleration. With this method, the issue that
the ego car velocity springs from 0 to 50 mph was corrected. The method
also detects another car on the lane using "getVehicleAhead" and changes
automatically to the next lane (still not looking before if there are
cars there).

The method "getVehicleAhead" searchs within the sensor fusion vector if
there are cars ahead on its current lane. It was tested together with
"changeTrajectory". If another vehicle is detected in a distance smaller
than the expected traveled distance after this part of the trajectory,
true is returned, otherwise false is returned. Also the closest vehicle
is passed as a reference.

The method "getVehicleBehind" works similarly as "getVehicleAhead"
detecting the cars with position s smaller than the position of ego
vehicle and returning true if a car is detected and false otherwise. The
closest vehicle is also passed as a reference.

The method "getKinematicsOfLane" gets the average velocity and
acceleration of a given lane, it is still not tested but it will be used
to get the new velocity and acceleration of all the lanes and determine
which is better to drive in.

The method "keepLineTraj" gets the velocity and acceleration of the ego
vehicle's current lane, it will be used to determine if staying at that
lane is the best way to go further or if it better to take another lane.