Methods "getKinematicsOfLane" and "keepLaneTraj" tested
The method "getKinematicsOfLane" was tested, changed because the results
were not as expected. This method returns the velocity and acceleration
of a lane based on the position and velocity of the vehicles ahead the
ego vehicle on that lane. If the detected vehicle is too close to the
ego vehicle, velocity will be given by the distance between the two
vehicles minus a distance buffer added to the current velocity of the
vehicle ahead minus the expected change on velocity of the ego car given
by 0.5*its acceleration. This velocity is capped by the highest possible
velocity the car can generate with its maximum acceleration and by the
target speed which corresponds to the speed limit of the road. If no car
is ahead, the minumum speed between the target speed and the maximum
possible speed is selected as the new velocity of the car. The car
acceleration is given by the difference between the new velocity and the
current velocity.

The method "keepLaneTraj" was tested also being used on
"changeTrajectory" to update the velocity and acceleration of the ego
car. It works well keeping the car with a relatively constant speed and
far away from the car ahead on the lane.