Methods "prepLaneChangeTraj" and "laneChangeTraj" programmed and tested.
Both methods are used in order to generate a lane change.

- prepLaneChangeTraj changes the speed of the ego vehicle to the minimum
between the current lane and the next lane. It was tested using it
within "changeTrajectory" to get the new velocity and acceleration
values from it. As expected, the velocity and acceleration decreases if
the desired lane is going slower than the current lane, but also as
desired, that only happens if there is no vehicle behind the ego vehicle
at that lane now.
- lanechangeTraj is the method which actually changes the lane. For now,
it only looks if there is a car between the current s of the ego vehicle
and the expected s after this part of the trajectory to say that a lane
change is safe. In order to make it safer it can also be looked for a
car behind the ego vehicle on that lane. It was tested like
"prepLaneChangeTraj" calling it from "changeTrajectory" at every cycle
which caused the software to change the lane always when a car is not at
the expected s value on that lane.