Methods "generateStateTraj" and "possibleNextStates" added.
These two methods will be used in order to select the best next state.
"generateStateTraj" executes "keepLaneTraj", "prepLaneChangeTraj" or
"laneChangeTraj" based on state. It was tested replacing it by the
respective method  on the tests of the other three methods.

The method "possibleNextStates" returns a vector of strings indicating
the possible next states based on the current state of the FSM. It was
tested by changing the current state of the ego vehicle and printing the
possible new states, observing the expected behaviour.

Also the method "SplineTraj" was changed. Now it is checked before the
interpolation if there are duplicated points, and if that is the case,
one of the points is deleted in order to avoid crashing from the spline
class.