Cost functions normalized
All cost functions were normalized so they return similar numbers. Doing
this the ego vehicle became more dinamic on its movements, taking more
attention to the speed and position of the cars closer on their lanes.
Also the weights were tunned and a new cost function
"laneChangeWhenNextCarCloseCost" was added but not used. This function
should prevent the ego vehicle to change lanes when it is too close to
the next vehicle on its lane.

Also the code was cleaned and commented. A video of 15 of driving
without problems was recorded.